// web/src/app/app/interfaces.ts
export enum ChatSessionSharedStatus {
  Private = "private",
  Public = "public",
}
// ...
export interface ChatSession {
  id: string;
  name: string;
  persona_id: number;
  time_created: string;
  time_updated: string;
  shared_status: ChatSessionSharedStatus;
  project_id: number | null;
  current_alternate_model: string;
  current_temperature_override: number | null;
  current_reasoning_effort_override: ReasoningEffortOverride | null;
}